# Conflicts:
#	app/helpers.php
#	public/assets/css/custom.css

+++ dbv-3/data/schema/tblVendorCDR.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `tblVendorCDR` (
-  `VendorCDRID` int(11) NOT NULL AUTO_INCREMENT,
+  `VendorCDRID` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
   `VendorCDRHeaderID` int(11) NOT NULL,
   `connect_time` datetime DEFAULT NULL,
   `disconnect_time` datetime DEFAULT NULL,
